services/load-generator/requirements.txt
paho-mqtt
kafka-python
requests
python-dotenv